pulled changes form upstream repo

+++ src/Console/ClearCommand.php
@@ -6,8 +6,8 @@ use Illuminate\Console\Command;
 use Illuminate\Console\ConfirmableTrait;
 use Illuminate\Queue\QueueManager;
 use Illuminate\Support\Arr;
+use Laravel\Horizon\Contracts\JobRepository;
 use Laravel\Horizon\RedisQueue;
-use Laravel\Horizon\Repositories\RedisJobRepository;
 
 class ClearCommand extends Command
 {
@@ -34,7 +34,7 @@ class ClearCommand extends Command
      *
      * @return int|null
      */
-    public function handle(RedisJobRepository $jobRepository, QueueManager $manager)
+    public function handle(JobRepository $jobRepository, QueueManager $manager)
     {
         if (! $this->confirmToProceed()) {
             return 1;
@@ -48,7 +48,9 @@ class ClearCommand extends Command
 
         $connection = Arr::first($this->laravel['config']->get('horizon.defaults'))['connection'] ?? 'redis';
 
-        $jobRepository->purge($queue = $this->getQueue($connection));
+        if (method_exists($jobRepository, 'purge')) {
+            $jobRepository->purge($queue = $this->getQueue($connection));
+        }
 
         $count = $manager->connection($connection)->clear($queue);
